Aggiunta Paypal
aggiunta paypal e ciò che serve per l'utilizzo

+++ src/main/java/gabrielebelluco/EnergyRoad/repositories/ProductAndServiceRepository.java
@@ -4,6 +4,7 @@ import gabrielebelluco.EnergyRoad.entities.ProductAndService;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -17,14 +18,13 @@ public interface ProductAndServiceRepository extends JpaRepository<ProductAndSer
 
     boolean existsByName(String name);
 
-
     List<ProductAndService> findByDescription(String description);
 
-    List<ProductAndService> findByPrice(double price);
+    List<ProductAndService> findByPrice(BigDecimal price);
 
     Optional<ProductAndService> findByName(String name);
 
-    List<ProductAndService> findByPriceLessThanEqual(Double maxPrice);
+    List<ProductAndService> findByPriceLessThanEqual(BigDecimal maxPrice);
 
-    List<ProductAndService> findByPriceGreaterThanEqual(Double minPrice);
+    List<ProductAndService> findByPriceGreaterThanEqual(BigDecimal minPrice);
 }
